+++ tests/cases/reject_cls_extends_nonexported.ts
@@ -1,0 +1,7 @@
+class Hidden {
+  x: number = 1;
+}
+
+export class Shown extends Hidden {
+  y: number = 2;
+}