+++ tests/cases/reject_cls_super_args.ts
@@ -1,0 +1,13 @@
+class Base {
+  x: number;
+
+  constructor(x: number) {
+    this.x = x;
+  }
+}
+
+class Derived extends Base {
+  constructor() {
+    super("one");
+  }
+}