+++ tests/cases/reject_cls_this_before_super.ts
@@ -1,0 +1,15 @@
+class Base {
+  x: number;
+
+  constructor(x: number) {
+    this.x = x;
+  }
+}
+
+class Derived extends Base {
+  y: number = 2;
+
+  constructor() {
+    super(this.y);
+  }
+}